+++ src/app/(view)/(auth)/sign-up/redirect/page.tsx
@@ -1,85 +1,28 @@
 "use client";
 
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import RegisterSetup from "@/_components/auth/register/RegisterSetup";
-import ConfirmModal from "@/_common/ConfirmModal";
-import TopHeader from "@/_common/TopHeader";
-import useFunnel from "@/_utils/hooks/useFunnel";
+import { useSearchParams, useRouter } from "next/navigation";
+import { Suspense, useEffect } from "react";
+import LoginLoading from "@/_components/auth/login/LoginLoading";
+import { useRegisterStore } from "@/_store/register";
 
-const STEPS = ["약관동의", "닉네임", "가입정보"];
-
-export default function Page() {
-  const { Funnel, currentStep, setStep } = useFunnel(STEPS[0]);
+function RegisterHandlerComponent() {
   const router = useRouter();
-  const step = STEPS.indexOf(currentStep) + 1;
-  const [registerCancelModalOpen, setRegisterCancelModalOpen] =
-    useState<boolean>(false);
-
-  const handleRegisterCancelModalOpen = () => {
-    setRegisterCancelModalOpen(true);
-  };
-
-  const handleRegisterCancelModalClose = () => {
-    setRegisterCancelModalOpen(false);
-  };
-
-  const handleStepChange = (newStep: string) => {
-    if (STEPS.includes(newStep)) {
-      setStep(newStep);
-    }
-  };
-
-  const handleRegisterCancel = () => {
-    router.replace("/");
-  };
-
-  const handleBack = () => {
-    const currentIndex = STEPS.indexOf(currentStep);
-    if (currentIndex > 0) {
-      setStep(STEPS[currentIndex - 1]);
-    } else {
-      setStep(STEPS[0]);
-    }
-  };
+  const searchParams = useSearchParams();
+  const { setEmail } = useRegisterStore();
 
   useEffect(() => {
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      event.preventDefault();
-    };
-
-    window.addEventListener("beforeunload", handleBeforeUnload);
-
-    return () => {
-      window.removeEventListener("beforeunload", handleBeforeUnload);
-    };
+    const email = searchParams.get("email");
+    setEmail(email ?? "");
+    router.replace("/sign-up");
   }, []);
 
+  return null;
+}
+
+export default function Page() {
   return (
-    <>
-      <TopHeader
-        title="회원가입"
-        step={step}
-        rest={STEPS.length - step}
-        goBack={() => handleBack()}
-        onClick={handleRegisterCancelModalOpen}
-      />
-      <RegisterSetup
-        steps={STEPS}
-        onNext={handleStepChange}
-        Funnel={Funnel}
-        Step={Funnel.Step}
-      />
-      {registerCancelModalOpen && (
-        <ConfirmModal
-          modalTitle="회원가입을 중단하시겠어요?"
-          modalDescription="마지막 단계에요! 그래도 중단하시겠어요?"
-          confirmText="중단하기"
-          cancelText="닫기"
-          handleConfirm={handleRegisterCancel}
-          handleCancel={handleRegisterCancelModalClose}
-        />
-      )}
-    </>
+    <Suspense fallback={<LoginLoading />}>
+      <RegisterHandlerComponent />
+    </Suspense>
   );
 }
